refactor: move GitHub account linking and settings to dedicated /profile/settings page
The GitHub OAuth connection UI was nested inside the profile edit form,
causing invalid HTML (nested forms) and poor discoverability. This moves
it to a new /profile/settings route alongside account settings (password,
delete account, data export), adds a Settings link to the user dropdown,
and simplifies the profile edit page to only handle profile data.
