+++ src/routes/profile/edit/+page.server.js
@@ -1,55 +1,8 @@
 import { redirect, fail } from '@sveltejs/kit';
 import { uploadImageAndReturnUrl, removeImage } from '$lib/server/service/imageUploadService.js';
 import { profileSchema } from '$lib/server/validator/profileScheme.js';
-import { deleteAccount } from '$lib/server/service/authUserService.js';
-import { getConnectionStatus, unlinkGitHub } from '$lib/server/service/githubConnectionService.js';
-import { linkIdentity } from '$lib/server/providers/authProvider.js';
-
-export async function load({ locals }) {
-  const user = locals.authUser;
-  if (!user) return { githubConnection: null };
-
-  const githubConnection = await getConnectionStatus(user.id, locals.supabase);
-  return { githubConnection };
-}
 
 export const actions = {
-  connectGitHub: async ({ locals, url }) => {
-    try {
-      const { url: authUrl } = await linkIdentity(locals.supabase, {
-        provider: 'github',
-        redirectTo: `${url.origin}/auth/github/callback`,
-        scopes: 'repo read:user',
-      });
-
-      if (authUrl) {
-        throw redirect(303, authUrl);
-      }
-
-      return fail(500, { error: 'Failed to get GitHub authorization URL.' });
-    } catch (err) {
-      if (err.status === 303) throw err;
-      return fail(500, { error: 'Failed to initiate GitHub connection.' });
-    }
-  },
-
-  disconnectGitHub: async ({ locals }) => {
-    const user = locals.authUser;
-    const supabase = locals.supabase;
-
-    if (!user) {
-      return fail(401, { error: 'Not authenticated' });
-    }
-
-    try {
-      await unlinkGitHub(user.id, supabase);
-    } catch (_) {
-      return fail(500, { error: 'Failed to disconnect GitHub account.' });
-    }
-
-    throw redirect(303, '/profile/edit');
-  },
-
   updateProfile: async ({ request, locals, fetch }) => {
     let supabase = locals.supabase;
 
@@ -99,32 +52,5 @@ export const actions = {
     }
 
     redirect(307, '/profile');
-  },
-
-  deleteAccount: async ({ cookies, locals }) => {
-    const user = locals.authUser;
-    const supabase = locals.supabase;
-
-    if (!user) {
-      return fail(401, { error: 'Not authenticated' });
-    }
-
-    try {
-      await deleteAccount(user.id, supabase);
-
-      for (const token of ['access_token', 'refresh_token']) {
-        cookies.set(token, '', {
-          path: '/',
-          expires: new Date(0),
-          httpOnly: true,
-          secure: true,
-          sameSite: 'strict',
-        });
-      }
-    } catch (_) {
-      return fail(500, { error: 'Failed to delete account. Please try again later.' });
-    }
-
-    redirect(307, '/');
   },
 };
