+++ src/routes/profile/edit/+page.svelte
@@ -1,28 +1,13 @@
 <script>
   import ProfileForm from '$lib/ProfileForm.svelte';
   import ProfileLinks from '$lib/ProfileLinks.svelte';
-  import Settings from '$lib/Settings.svelte';
-  import Interests from '$lib/Interests.svelte';
   import Icon from '@iconify/svelte';
   import { applyAction, enhance } from '$app/forms';
   import { toast } from 'svelte-sonner';
-  import { page } from '$app/stores';
-  import { onMount } from 'svelte';
 
-  let activeNavItem = 'Profile';
   let loading = false;
   export let data;
   let user = data.user;
-  let githubConnection = data.githubConnection;
-
-  onMount(() => {
-    const urlParams = $page.url.searchParams;
-    if (urlParams.get('github') === 'linked') {
-      toast.success('GitHub account connected successfully');
-    } else if (urlParams.get('error') === 'github_link_failed') {
-      toast.error('Failed to connect GitHub account. Please try again.');
-    }
-  });
 </script>
 
 <!-- Main Content Container -->
@@ -74,7 +59,7 @@
 
           <!-- Right Column - Links & Social -->
           <div class="space-y-8">
-            <ProfileLinks {user} {githubConnection} />
+            <ProfileLinks {user} />
           </div>
         </div>
